+++ helpers/Html.php
@@ -8,7 +8,7 @@
 
 namespace humhub\modules\custom_pages\helpers;
 
-use humhub\libs\Html as BaseHtml;
+use humhub\helpers\Html as BaseHtml;
 
 class Html extends BaseHtml
 {