+++ modules/template/views/admin/importSource.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) HumHub GmbH & Co. KG
@@ -6,35 +7,19 @@
  */
 
 use humhub\modules\custom_pages\modules\template\models\forms\ImportForm;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\modules\ui\view\components\View;
-use humhub\widgets\Button;
-use humhub\widgets\ModalButton;
-use humhub\widgets\ModalDialog;
+use humhub\widgets\modal\Modal;
+use humhub\widgets\modal\ModalButton;
 
-/* @var $this View */
 /* @var $model ImportForm */
 ?>
-
-<?php ModalDialog::begin([
-    'header' => Yii::t('CustomPagesModule.template', '<strong>Import</strong> Template'),
+<?php $form = Modal::beginFormDialog([
+    'title' => Yii::t('CustomPagesModule.template', '<strong>Import</strong> Template'),
+    'footer' => ModalButton::cancel() .
+        ModalButton::primary(Yii::t('CustomPagesModule.template', 'Import'))->submit(),
+    'form' => ['options' => ['enctype' => 'multipart/form-data']],
 ]) ?>
-
-<?php $form = ActiveForm::begin(['options' => ['enctype' => 'multipart/form-data']]) ?>
-
-<div class="modal-body">
     <div class="alert alert-warning">
         <?= Yii::t('CustomPagesModule.template', 'If a template with the same name already exists, it will be replaced with the data from your import file.') ?>
     </div>
-
     <?= $form->field($model, 'file')->fileInput() ?>
-</div>
-
-<div class="modal-footer">
-    <?= ModalButton::cancel() ?>
-    <?= Button::primary(Yii::t('CustomPagesModule.template', 'Import'))->submit() ?>
-</div>
-
-<?php ActiveForm::end() ?>
-
-<?php ModalDialog::end() ?>
+<?php Modal::endFormDialog() ?>