+++ modules/template/views/admin/index.php
@@ -6,14 +6,14 @@
  * @license https://www.humhub.com/licences
  */
 
-use humhub\libs\Html;
-use humhub\modules\custom_pages\modules\template\assets\TemplatePageStyleAsset;
+use humhub\helpers\Html;
+use humhub\modules\custom_pages\modules\template\assets\TemplateAsset;
 use humhub\modules\custom_pages\modules\template\components\TemplateActionColumn;
 use humhub\modules\custom_pages\modules\template\models\Template;
 use humhub\modules\custom_pages\modules\template\models\TemplateSearch;
 use humhub\modules\custom_pages\widgets\AdminMenu;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 use humhub\widgets\GridView;
 use yii\data\ActiveDataProvider;
 use yii\helpers\Url;
@@ -21,13 +21,13 @@ use yii\helpers\Url;
 /* @var ActiveDataProvider $dataProvider */
 /* @var TemplateSearch $searchModel */
 
-TemplatePageStyleAsset::register($this);
+TemplateAsset::register($this);
 ?>
 <div class="panel panel-default">
     <div class="panel-heading"><?= Yii::t('CustomPagesModule.base', '<strong>Custom</strong> Pages') ?></div>
     <?= AdminMenu::widget() ?>
     <div class="panel-body">
-        <div class="help-block">
+        <div class="text-body-secondary">
             <?= Yii::t('CustomPagesModule.base', 'Manage layouts, snippet layouts, and containers. Layouts define page structures, snippet layouts are used in sidebars or sections, and containers are reusable content blocks.') ?>
         </div>
     </div>
@@ -40,7 +40,7 @@ TemplatePageStyleAsset::register($this);
                     <?= $form->field($searchModel, 'name')
                         ->textInput(['placeholder' => Yii::t('CustomPagesModule.template', 'Search by template ID or name')])
                         ->label(false) ?>
-                    <?= Button::defaultType()->icon('search')->submit() ?>
+                    <?= Button::light()->icon('search')->submit() ?>
                 </div>
                 <?= $form->field($searchModel, 'type')
                     ->dropDownList(['' => Yii::t('CustomPagesModule.template', 'Type (All)')] + $searchModel->getTypeOptions())
@@ -51,7 +51,7 @@ TemplatePageStyleAsset::register($this);
                 <?php ActiveForm::end() ?>
             </div>
             <div>
-                <?= Button::info(Yii::t('CustomPagesModule.base', 'Import'))
+                <?= Button::accent(Yii::t('CustomPagesModule.base', 'Import'))
                     ->action('ui.modal.load', ['import-source'])
                     ->icon('download')
                     ->style('margin-right:5px') ?>