+++ modules/template/widgets/views/templateElementAdminRow.php
@@ -1,6 +1,7 @@
 <?php
 
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Badge;
+use humhub\widgets\bootstrap\Button;
 use yii\helpers\Html;
 
 /* @var $model humhub\modules\custom_pages\modules\template\models\TemplateElement */
@@ -10,25 +11,19 @@ use yii\helpers\Html;
         #<strong><?= Html::encode($model->name) ?> </strong>
     </td>
     <td>
-        <small>
-            <span class="label label-success"><?= $model->getLabel() ?></span>
-        </small>
+        <?= Badge::success($model->getLabel()) ?>
         <?php if (!$model->hasDefaultContent()) : ?>
-            <small>
-                <span class="label label-warning"><?= Yii::t('CustomPagesModule.base', 'Empty') ?></span>
-            </small>
+            <?= Badge::warning(Yii::t('CustomPagesModule.base', 'Empty')) ?>
         <?php else: ?>
-            <small>
-                <span class="label btn-success"><?= Yii::t('CustomPagesModule.base', 'Default') ?></span>
-            </small>
+            <?= Badge::success(Yii::t('CustomPagesModule.base', 'Default')) ?>
         <?php endif; ?>
     </td>
 
     <td>
     <?php if ($model->template->canEdit()) : ?>
-        <?= Button::primary()->icon('pencil')->xs()
+        <?= Button::primary()->icon('pencil')->sm()
             ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
-        <?= Button::danger()->icon('times')->xs()
+        <?= Button::danger()->icon('times')->sm()
             ->action('deleteElementSubmit', ['/custom_pages/template/admin/delete-element', 'id' => $model->id])
             ->confirm(
                 Yii::t('CustomPagesModule.template', '<strong>Confirm</strong> element deletion'),
@@ -36,7 +31,7 @@ use yii\helpers\Html;
                 Yii::t('CustomPagesModule.base', 'Delete'),
             ) ?>
     <?php else : ?>
-        <?= Button::info()->icon('eye')->xs()
+        <?= Button::accent()->icon('eye')->sm()
             ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
     <?php endif; ?>
     </td>