+++ modules/template/widgets/views/templateMenu.php
@@ -5,23 +5,22 @@
  * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\ui\menu\MenuLink;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
 
 /* @var MenuLink[] $entries */
 /* @var array $options */
 ?>
 <?= Html::beginTag('div', $options) ?>
-<?= Button::defaultType()
+<?= Button::light()
     ->icon('ellipsis-h')
-    ->cssClass('dropdown-toggle')
-    ->options(['data-toggle' => 'dropdown'])
+    ->options(['data-bs-toggle' => 'dropdown'])
     ->loader(false) ?>
-<ul class="dropdown-menu pull-right">
+<ul class="dropdown-menu">
     <?php foreach ($entries as $entry) : ?>
         <li>
-            <?= $entry->render() ?>
+            <?= $entry->render(['class' => 'dropdown-item']) ?>
         </li>
     <?php endforeach; ?>
 </ul>