+++ views/common/list.php
@@ -18,7 +18,7 @@ use humhub\modules\custom_pages\widgets\TargetPageList;
     <div class="panel-body">
         <div class="clearfix">
             <h4><?= Yii::t('CustomPagesModule.base', 'Overview') ?></h4>
-            <div class="help-block">
+            <div class="text-body-secondary">
                 <?= Yii::t('CustomPagesModule.view', 'This page lists all available custom content entries.'); ?>
             </div>
         </div>